renders: remove reconfigure nolerp scrap textures

+++ src/client/refresh/files/stb.c
@@ -1105,7 +1105,7 @@ R_FloodFillSkin(byte *skin, int skinwidth, int skinheight, const unsigned *table
 }
 
 unsigned *
-R_Convert8to32(const byte *data, int width, int height, const unsigned *table_8to24)
+R_Convert8to32(const byte *data, size_t width, size_t height, const unsigned *table_8to24)
 {
 	unsigned *trans;
 	size_t i, s;
